size_t no matching function for call to min
git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@64128 954022d7-b5bf-4e40-9824-e11837661b57

+++ ASCOfficeOdfFile/src/docx/docx_conversion_context.cpp
@@ -423,7 +423,7 @@ mc:Ignorable=\"w14 wp14\">";
         strm << L"<w:abstractNum w:abstractNumId=\"" << abstractNumId << "\">";
         numIds.push_back(abstractNumId);		
 		
-		for (int i = 0; i < (std::min)( content.size(), (unsigned int)9); i++)
+		for (int i = 0; i < (std::min)( content.size(), (size_t)9); i++)
         {
             start_text_list_style(inst->get_text_list_style()->get_style_name());
             content[i]->docx_convert(*this);
